Test UI with espresso

+++ app/src/main/java/bakingtime/maxgaj/udacity/com/bakingtime/Model/Ingredient.java
@@ -8,7 +8,7 @@ public class Ingredient implements Parcelable {
     private String measure;
     private String ingredient;
 
-    public Ingredient(long quantity, String measure, String ingredient){
+    public Ingredient(double quantity, String measure, String ingredient){
         this.quantity = quantity;
         this.measure = measure;
         this.ingredient = ingredient;
